Update clustering icons (#50)
* Update clustering icons

1. Profile icons are now clustered and they are shown on top of AP icons.
2. Profile icon clusters show a sample of contained pfps.
3. AP icons now indicate the verified status of contained APs.

* Animate map movement

* Clarify CenterZoomController constants

In passing I fixed the application of the zoom velocity value. Now a
higher velocity results in a faster animation as documented. This has no
affect on behaviour since we were using a value of 1 which is neutral.

* Document AccessPointCluster class and its combine method

In passing I realised that `covariant` is not needed on the combine
method parameter, it is only necessary on the superclass declaration,
and removed it.

* Use MapController only to animate map movement

This avoids the need to have a child widget of the map which passes the
map state to the MapCubit.