Map tests (#54)
* Add LocationRepository test

* Refactor map_flutter_map for readability

* Make map buttons react properly to their respective states

* Add tests for filter_map_button

* Add tests for map filter and location buttons

In passing I have partially fixed the existing phone screen test. Now it
needs to stub the page navigation and the check for the pinput should be
moved to a separate test on the relevant page.

* Fix sms verification page test

It now checks that we navigated rather than doing the navigation and
testing whether the PinInput is visible. Checking whether the PinInput
is visible should be done on the page which is being navigated to.

* Add map filter dialog tests

* Add MapFilterController test

* Add MapService tests

* Add LocationRepository test

* Refactor map_flutter_map for readability

* Make map buttons react properly to their respective states

* Add tests for filter_map_button

* Add tests for map filter and location buttons

In passing I have partially fixed the existing phone screen test. Now it
needs to stub the page navigation and the check for the pinput should be
moved to a separate test on the relevant page.

* Fix sms verification page test

It now checks that we navigated rather than doing the navigation and
testing whether the PinInput is visible. Checking whether the PinInput
is visible should be done on the page which is being navigated to.

* Add map filter dialog tests

* Add MapFilterController test

* Add MapService tests

* Make MapFilter wait for the sharedprefs to be ready before emitting its first value

* Fix flutter analyze issues from flutter_lints 2.x

* Remove code coverage job from workflow

Co-authored-by: bifrostyyy <bifrost@verifi.world>
Co-authored-by: Bifrost <103068176+bifrostyyy@users.noreply.github.com>